Various fixes and improvements suggested by SonarQube.
Fixes include:
- don't refer to subclass from parent class initialization.
- cast ints to long before assignment to avoid overflow
- Iterator.next() method should throw NoSuchElementException
- if getter is synchronized, setter should also be (and vice versa)
- if base method is synchronized, override should be as well
- unary minus on compareTo() result can cause problems if result is MIN_VALUE
- group regex to make | alternatives clearer
- don't use several overrides for compareTo() (see RelationInfo)

+++ engine/src/main/java/nl/inl/blacklab/codec/BlackLab40StoredFieldsReader.java
@@ -302,7 +302,7 @@ public class BlackLab40StoredFieldsReader extends BlackLabStoredFieldsReader {
                     int blockStartOffset = findBlockStartOffset(blockIndexOffset, blocksOffset, firstBlockNeeded);
 
                     // Try to make sure we have a large enough buffer available
-                    long decodeBufferLengthLong = valueLengthChar * UTF8_MAX_BYTES_PER_CHAR + ESTIMATED_DECODE_OVERHEAD;
+                    long decodeBufferLengthLong = (long) valueLengthChar * UTF8_MAX_BYTES_PER_CHAR + ESTIMATED_DECODE_OVERHEAD;
                     if (decodeBufferLengthLong > Constants.JAVA_MAX_ARRAY_SIZE)
                         decodeBufferLengthLong = Constants.JAVA_MAX_ARRAY_SIZE;
                     final int decodeBufferLength = (int) decodeBufferLengthLong;
