+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansNGrams.java
@@ -115,7 +115,7 @@ class SpansNGrams extends BLSpans {
                 return NO_MORE_DOCS; // no more docs; we're done
             }
             // Get document length and reset currentStart/currentEnd so we can check if there's actually hits
-            currentDocLength = lengthGetter.getFieldLength(currentDoc) - BlackLabIndexAbstract.IGNORE_EXTRA_CLOSING_TOKEN;
+            currentDocLength = (long)lengthGetter.getFieldLength(currentDoc) - BlackLabIndexAbstract.IGNORE_EXTRA_CLOSING_TOKEN;
             currentStart = currentEnd = -1;
         } while (currentDocLength < min || nextStartPosition() == NO_MORE_POSITIONS);
         atFirstInCurrentDoc = true;
